pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[project]
name = "wfp-flask-template"
version = "0.0.1"
description = "Waterfall Project Flask Template"
requires-python = ">=3.9"
dependencies = [
    "colorlog",
    "structlog",
    "python-dotenv",
    "Flask",
    "flask-cors",
    "flask-marshmallow",
    "Flask-Migrate",
    "Flask-Limiter",
    "flask-restful",
    "Flask-SQLAlchemy",
    "marshmallow-sqlalchemy",
    "marshmallow>=4.1.2",
    "psycopg2-binary",
    "PyJWT",
    "gunicorn",
    "requests",
    "prometheus-flask-exporter",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "ruff",
    "mypy",
    "bandit",
    "pre-commit",
    "interrogate",
    "genbadge[coverage]",
    "types-requests",
]

[tool.pytest.ini_options]
# Add current directory to Python path automatically
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# Ignore certain directories
norecursedirs = [".git", ".venv", "venv", "__pycache__", "*.egg-info"]
# Show extra test summary info
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
# Register custom markers
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Black formatter configuration
[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312", "py313"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | venv
  | __pycache__
  | \.pytest_cache
  | \.mypy_cache
  | \.ruff_cache
  | build
  | dist
  | migrations
)/
'''

# isort configuration (import sorting)
[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*", "*/__pycache__/*", ".venv/*", "venv/*"]
known_first_party = ["app"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

# Ruff linter configuration (replaces flake8, pylint, etc.)
[tool.ruff]
line-length = 88
target-version = "py39"
extend-exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    "build",
    "dist",
    "migrations",
]

[tool.ruff.lint]
# Enable pycodestyle (E, W), pyflakes (F), isort (I), and more
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "D",      # pydocstyle (docstring conventions)
]
ignore = [
    "E501",   # Line too long (handled by black)
    "B008",   # Do not perform function calls in argument defaults
    "W191",   # Indentation contains tabs
    "B904",   # Allow raising exceptions without from e
    "D203",   # 1 blank line required before class docstring (incompatible with D211)
    "D213",   # Multi-line docstring summary should start at the second line (incompatible with D212)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/*" = ["S101", "D"]  # Allow assert in tests, no docstring requirement
"scripts/*" = ["D"]         # No docstring requirement for scripts

[tool.ruff.lint.pydocstyle]
convention = "google"  # Use Google-style docstrings

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# MyPy type checking configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start with false, enable gradually
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
strict_optional = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false

# Bandit security linter configuration
[tool.bandit]
exclude_dirs = ["/tests", "/migrations", "/.venv", "/venv"]
skips = ["B101"]  # Skip assert_used in tests

# Interrogate docstring coverage configuration
[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-magic = true
ignore-semiprivate = true
ignore-private = true
ignore-property-decorators = true
ignore-module = false
ignore-nested-functions = true
ignore-nested-classes = true
fail-under = 80  # Minimum 80% docstring coverage
exclude = ["tests", "scripts", "migrations", ".venv", "venv"]
verbose = 1
quiet = false
whitelist-regex = []
color = true
omit-covered-files = false
generate-badge = "docs/assets/"
badge-format = "svg"
